# openapi/user-storage-bff.yaml
openapi: 3.1.0
info:
  title: KYC Backend BFF Flow API
  version: 0.2.2
  description: >
    BFF-facing v2 flow runtime API. Only flow sessions, flows, steps, user reads, and upload endpoints remain on this surface.
servers:
  - url: https://kyc-internal.example.local
...
security:
  - KcSignature: [ ]
paths:
...
  /uploads/presign:
    post:
      tags: [ uploads ]
      summary: Create presigned upload target
      operationId: presignUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignUploadRequest"
      responses:
        "200":
          description: Presigned upload response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignUploadResponse"
  /uploads/complete:
    post:
      tags: [ uploads ]
      summary: Complete upload registration
      operationId: completeUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteUploadRequest"
      responses:
        "200":
          description: Evidence registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceRef"
...
components:
...
  schemas:
...
    UploadMethod:
      type: string
      enum: [ PUT ]
    UploadEncryptionMode:
      type: string
      enum: [ SSE_S3, SSE_KMS, SSE_C ]
    PresignUploadRequest:
      type: object
      additionalProperties: false
      properties:
        sessionId: { type: string }
        stepId: { type: string }
        userId: { type: string }
        purpose: { type: string }
        assetType: { type: string }
        mime: { type: string }
        sizeBytes: { type: integer, minimum: 1 }
        encryption:
          type: object
          additionalProperties: false
          properties:
            mode:
              $ref: "#/components/schemas/UploadEncryptionMode"
          required: [ mode ]
      required: [ sessionId, stepId, userId, purpose, assetType, mime, sizeBytes ]
    PresignUploadResponse:
      type: object
      additionalProperties: false
      properties:
        uploadId: { type: string }
        bucket: { type: string }
        objectKey: { type: string }
        method:
          $ref: "#/components/schemas/UploadMethod"
        url: { type: string }
        headers:
          type: object
          additionalProperties: { type: string }
        expiresAt: { type: string, format: date-time }
      required: [ uploadId, bucket, objectKey, method, url, headers, expiresAt ]
    CompleteUploadRequest:
      type: object
      additionalProperties: false
      properties:
        sessionId: { type: string }
        stepId: { type: string }
        uploadId: { type: string }
        bucket: { type: string }
        objectKey: { type: string }
        etag: { type: string, nullable: true }
        computedSha256: { type: string, nullable: true }
      required: [ sessionId, stepId, uploadId, bucket, objectKey ]
    EvidenceRef:
      type: object
      additionalProperties: false
      properties:
        evidenceId: { type: string }
        sessionId: { type: string }
        stepId: { type: string }
        assetType: { type: string }
        sha256: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [ evidenceId, sessionId, stepId, assetType, createdAt ]
...
  securitySchemes:
    KcSignature:
      type: apiKey
      in: header
      name: x-kc-signature